Bootstrap 5 Buttons - Active/Disabled Buttons
A button can be set to an active (appear pressed) or a disabled (unclickable) state.

The class .active makes a button appear pressed, and the disabled attribute makes a button unclickable. Note that <a> elements do not support the disabled attribute and must therefore use the .disabled class to make it visually appear disabled.