CSS Progress Bars - Basic Progress Bar
A progress bar can be used to show how far a user is in a process.

To create a default progress bar, add a .progress class to a container element and add the .progress-bar class to its child element. Use the CSS width property to set the width of the progress bar