BS5 Navbars - Fixed Navigation Bar
A fixed navigation bar stays visible in a fixed position (top or bottom) independent of the page scroll.

The navigation bar can also be fixed at the top or at the bottom of the page.

A fixed navigation bar stays visible in a fixed position (top or bottom) independent of the page scroll.

The .fixed-top class makes the navigation bar fixed at the top