BS5 Toasts - How To Create A Toast
The toast component is like an alert box that is only shown for a couple of seconds when something happens (i.e. when the user clicks on a button, submits a form, etc.).

In this example, we use the .show class to always show the toast by default. You can close it by clicking on the close (x) icon inside the toast header.