Bootstrap 5 Offcanvas - Responsive Offcanvas Menu
You can also hide or show the offcanvas menu on different screen widths, with the .offcanvas-sm|md|lg|xl|xxl classes.

In this example, we hide the offcanvas menu on screens larger than 991px wide. Note that we also hide the button that opens the offcanvas (d-lg-none).

Resize the browser window to see the result.