Bootstrap 5 Grid Small - Small Grid
The following example will result in a 33.3%/66.6% split on small, medium, large and xlarge devices (576px and above). On extra small devices, it will stack (100% width).

Resize the browser window to see the effect.

This example will result in a 50%/50% split on small, medium, large and xlarge devices (576px and above). On extra small devices, it will stack (100% width).