Bootstrap 5 Grid Examples - Mix and Match
Resize the browser window to see the effect.

This example demonstrates a 50%/50% split on extra small devices and 75%/25% split on larger devices.

This example demonstrates a 58%/42% split on extra small, small and medium devices and 66.3%/33.3% split on large and xlarge devices.

his example demonstrates a 25%/75% split on small devices, a 50%/50% split on medium devices, and a 33%/66% split on large and xlarge devices. On extra small devices, it will automatically stack (100%).